agentic-ai/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
motor>=3.5
pydantic>=2.7
pydantic-settings>=2.4
python-dotenv>=1.0
openai>=1.43.0
python-jose[cryptography]>=3.3
email-validator>=2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1